Moved Lead batch limiter code to trait
